requirements.txt
mcp==1.1.0
httpx==0.28.1
python-dotenv==1.0.1
pydantic==2.10.3
typing-extensions==4.12.2
pytest==8.3.4
pytest-asyncio==0.25.0
fastapi==0.115.5
uvicorn==0.32.1